% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdata.R
\name{msdata}
\alias{msdata}
\title{msdata}
\usage{
msdata(x)
}
\arguments{
\item{x}{In quotes, the name of the data set, all caps.}
}
\value{
a data set
}
\description{
A function of brevity. It streamlines where to pull the MS Data sets from.
}
\examples{
\dontrun{msdata(x = "DDT")}
}
