% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBivarNorm.R
\name{myBivarNorm}
\alias{myBivarNorm}
\title{myBivarNorm}
\usage{
myBivarNorm(df, alpha = 0.05, contour = 0.5)
}
\arguments{
\item{df}{data frame}

\item{alpha}{alpha level for Shapiro-Wilk hypothesis test (univariate)}

\item{contour}{contour ellipse level}
}
\value{
Q-Q plots and histograms of both variables, a Chi-Sq Plot of ordered distances, and potential results of the question of Normality
}
\description{
A function to test a bivariate data set for Univariate and Bivariate Normality
}
\examples{
\dontrun{myBivarNorm(df = micro, alpha = 0.05, contour = 0.5)}
}
