% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myClusterF.R
\name{myClusterF}
\alias{myClusterF}
\title{myClusterF}
\usage{
myClusterF(df, Sim = FALSE)
}
\arguments{
\item{df}{data frame}

\item{Sim}{are the values similarities or differences}
}
\value{
a plot of three dendrograms for the three linkages and a list of the heights
}
\description{
A function which produces dendrogram clusterings for Single, Complete, and Average Linkages
}
\examples{
\dontrun{myClusterF(df = LANGUAGES)}
}
