% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFPC.R
\name{myFPC}
\alias{myFPC}
\title{myFPC}
\usage{
myFPC(sigma, m)
}
\arguments{
\item{sigma}{covariance matrix}

\item{m}{number of factors}
}
\value{
a list of factor loadings and cumulative variance accounted for by each factor
}
\description{
A function which will produce the factor analysis data for principal components and max likelihood methods
}
\examples{
\dontrun{myFPC(sigma = S, m = 4)}
}
