% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myJWdata.R
\name{myJWdata}
\alias{myJWdata}
\title{myJWdata}
\usage{
myJWdata(x)
}
\arguments{
\item{x}{In quotes, the name of the data set.}
}
\value{
a data set
}
\description{
A function of brevity. It streamlines where to pull the JW Data sets from.
}
\examples{
\dontrun{myJWdata(x = "P1-4")}
}
