% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myKmeans.R
\name{myKmeans}
\alias{myKmeans}
\title{myKmeans}
\usage{
myKmeans(df, k, x, y, z)
}
\arguments{
\item{df}{data frame}

\item{k}{number of clusters}

\item{x}{x variable}

\item{y}{y variable}

\item{z}{for table information}
}
\value{
a plot of the kmeans clusters and tables of counts for each cluster
}
\description{
myKmeans
}
\examples{
\dontrun{myKmeans(ddt, 2, LENGTH, WEIGHT, SPECIES)}
}
