% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPCA.R
\name{myPCA}
\alias{myPCA}
\title{myPCA}
\usage{
myPCA(df)
}
\arguments{
\item{df}{a numeric data frame}
}
\value{
a scree plot, the coefficients for the Primary Components, and cumulative variance explained by each PC
}
\description{
A function which calculates the coefficients for the Primary Components, cumulative variance explained by each PC, and produces a scree plot
}
\examples{
\dontrun{myPCA(df=logMALE)}
}
