% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myScreePlot.R
\name{myScreePlot}
\alias{myScreePlot}
\title{myScreePlot}
\usage{
myScreePlot(df)
}
\arguments{
\item{df}{data frame}
}
\value{
a scree plot of amount of variance explained by the principal components
}
\description{
myScreePlot
}
\examples{
\dontrun{myScreePlot(ddt)}
}
